+++ src/api/poller.ts
@@ -24,7 +24,7 @@ export async function getMoarData(): Promise<boolean> {
 
     const pollers = [
         MastoApi.instance.getRecentNotifications.bind(MastoApi.instance),
-        MastoApi.instance.getUserRecentToots.bind(MastoApi.instance),
+        MastoApi.instance.getRecentUserToots.bind(MastoApi.instance),
         // TODO: getRecentFavourites API doesn't use maxId argument BUT you can page as far back as you want
         // MastoApi.instance.getRecentFavourites.bind(MastoApi.instance),
     ];