+++ src/scorer/feature/most_replied_accounts_scorer.ts
@@ -17,7 +17,7 @@ export default class MostRepliedAccountsScorer extends FeatureScorer {
     // would require a lot of API calls, so it's just working with the account ID which is NOT
     // unique across all servers.
     async prepareScoreData(): Promise<StringNumberDict> {
-        const recentToots = await MastoApi.instance.getUserRecentToots();
+        const recentToots = await MastoApi.instance.getRecentUserToots();
         const recentReplies = recentToots.filter(toot => toot?.inReplyToAccountId);
         return countValues<Toot>(recentReplies, (toot) => toot?.inReplyToAccountId);
     };