+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
 * `MastoApi.instance.setBackgroundConcurrency()`, `Config.maxConcurrentRequestsBackground` option. Handle missing `Storage.getIdentity()` if possibl
+* Ignore retoots when computting users hashtag history
 
 ### v0.24.5
 * Remove blocked keywrods (no bueno) and followed tags (they show up on their own) from the list of tags to scan for new toots